+++ docs/create.py
@@ -128,8 +128,8 @@ In order to efficiently store and transfer data, PyPGx uses the ZIP archive file
       - Target gene.
       - ``CYP2D6``, ``GSTT1``
     * - ``Platform``
-      - NGS platform.
-      - ``WGS``, ``Targeted``
+      - Genotyping platform.
+      - ``WGS``, ``Targeted``, ``Chip``
     * - ``Program``
       - Name of the phasing program.
       - ``Beagle``
@@ -156,7 +156,7 @@ Notably, all archive files have defined semantic types, which allows us to ensur
     * Requires following metadata: ``Gene``, ``Assembly``, ``SemanticType``, ``Control``.
 - ``SampleTable[Alleles]``
     * TSV file for storing target gene's candidate star alleles for each sample.
-    * Requires following metadata: ``Gene``, ``Assembly``, ``SemanticType``, ``Program``.
+    * Requires following metadata: ``Platform``, ``Gene``, ``Assembly``, ``SemanticType``, ``Program``.
 - ``SampleTable[CNVCalls]``
     * TSV file for storing target gene's CNV call for each sample.
     * Requires following metadata: ``Gene``, ``Assembly``, ``SemanticType``, ``Control``.
@@ -174,13 +174,13 @@ Notably, all archive files have defined semantic types, which allows us to ensur
     * Requires following metadata: ``Control``, ``Assembly``, ``SemanticType``, ``Platform``.
 - ``VcfFrame[Consolidated]``
     * VcfFrame for storing target gene's consolidated variant data.
-    * Requires following metadata: ``Gene``, ``Assembly``, ``SemanticType``, ``Program``.
+    * Requires following metadata: ``Platform``, ``Gene``, ``Assembly``, ``SemanticType``, ``Program``.
 - ``VcfFrame[Imported]``
     * VcfFrame for storing target gene's raw variant data.
-    * Requires following metadata: ``Gene``, ``Assembly``, ``SemanticType``.
+    * Requires following metadata: ``Platform``, ``Gene``, ``Assembly``, ``SemanticType``.
 - ``VcfFrame[Phased]``
     * VcfFrame for storing target gene's phased variant data.
-    * Requires following metadata: ``Gene``, ``Assembly``, ``SemanticType``, ``Program``.
+    * Requires following metadata: ``Platform``, ``Gene``, ``Assembly``, ``SemanticType``, ``Program``.
 
 Getting help
 ============